# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AmbassadorChatInput {
  /**
     * @minLength 1
     * @maxLength 800
     */
  message: string;
}

export interface AmbassadorLink {
  label: string;
  href: string;
  section: string;
}

export interface AmbassadorChatResponse {
  answer: string;
  links: AmbassadorLink[];
}