# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: ambassador
    description: Evidence-grounded recruiter ambassador
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /ambassador/chat:
    post:
      operationId: askAmbassador
      tags: [ambassador]
      summary: Ask the virtual recruiter ambassador
      description: Answers recruiter questions using only verified resume content.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AmbassadorChatInput"
      responses:
        "200":
          description: Grounded ambassador response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AmbassadorChatResponse"
        "400":
          description: Invalid request
        "429":
          description: Rate limit exceeded
        "500":
          description: Upstream assistant unavailable
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AmbassadorChatInput:
      type: object
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 800
      required:
        - message
    AmbassadorLink:
      type: object
      properties:
        label:
          type: string
        href:
          type: string
        section:
          type: string
      required:
        - label
        - href
        - section
    AmbassadorChatResponse:
      type: object
      properties:
        answer:
          type: string
        links:
          type: array
          items:
            $ref: "#/components/schemas/AmbassadorLink"
      required:
        - answer
        - links