# lib/api-zod/src/generated/types/ambassadorChatInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AmbassadorChatInput {
  /**
     * @minLength 1
     * @maxLength 800
     */
  message: string;
}