# lib/api-zod/src/generated/types/ambassadorChatResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AmbassadorLink } from './ambassadorLink';

export interface AmbassadorChatResponse {
  answer: string;
  links: AmbassadorLink[];
}