# lib/api-zod/src/generated/types/ambassadorLink.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AmbassadorLink {
  label: string;
  href: string;
  section: string;
}